% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.Z.R
\name{cdf.Z}
\alias{cdf.Z}
\title{Evaluate the inverse Laplace-Stieltjes transform of a copula's generator}
\usage{
cdf.Z(cop, z)
}
\arguments{
\item{cop}{The copula}

\item{z}{Argument to the inverse Laplace-Stieltjes transform of the copula's generator}
}
\description{
\code{cdf.Z} evaluates the inverse Laplace-Stieltjes transform of the generator of the copula \code{cop} at \code{z}. Not: The evaluated mapping is a distribution function.
}
\keyword{internal}
