% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.Z.R
\name{sample.discrete}
\alias{sample.discrete}
\title{Generate a sample from a discrete distribution}
\usage{
sample.discrete(df, n)
}
\arguments{
\item{df}{The density function - It is assumed that the support is a subset of the natural numbers}

\item{n}{The desired sample size}
}
\description{
\code{sample.discrete} generates a sample of size \code{n} given its density function \code{df}
}
