% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.Z.R
\name{sample.Z}
\alias{sample.Z}
\title{Generate a sample from the inverse Laplace-Stieltjes transform of a copula's generator}
\usage{
sample.Z(cop, n)
}
\arguments{
\item{cop}{The copula}

\item{n}{The desired sample size}
}
\description{
\code{sample.Z} generates a sample of size \code{n} from the inverse Laplace-Stieltjes transform of the generator of the copula \code{cop}. For further details see \url{https://doi.org/10.1016/j.csda.2008.05.019} (especially table 1).
}
